analysis/requirements.txt
pandas
Pillow
matplotlib 